# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fondarium test tast api documentation",
        "description": "Fondarium test tast",
        "contact": {
            "email": "vladshvec7@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/autos": {
            "get": {
                "tags": [
                    "Autos"
                ],
                "summary": "Get all cars for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "indexCar",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Autos"
                ],
                "summary": "Update car",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateCar",
                "requestBody": {
                    "description": "Update car object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "auto_mark",
                                    "auto_number",
                                    "auto_color"
                                ],
                                "properties": {
                                    "auto_mark": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Volkswagen"
                                    },
                                    "auto_number": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456"
                                    },
                                    "auto_color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "DarkRed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Autos"
                ],
                "summary": "Create car",
                "description": "This can only be done by the logged in user.",
                "operationId": "createCar",
                "requestBody": {
                    "description": "Create car object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "auto_mark",
                                    "auto_number",
                                    "auto_color"
                                ],
                                "properties": {
                                    "auto_mark": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Volkswagen"
                                    },
                                    "auto_number": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456"
                                    },
                                    "auto_color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "DarkRed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/autos/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "This can only be done by the logged in user.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "User id",
                        "in": "path",
                        "description": "User id to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Per page count",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "This can only be done by the logged in user.",
                "operationId": "update",
                "requestBody": {
                    "description": "Update user object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "role_id",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "testmail@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456"
                                    },
                                    "role_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Andrey"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "create",
                "requestBody": {
                    "description": "Create user object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "role_id",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456"
                                    },
                                    "role_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Andrey"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "delete": {
                "tags": [
                    "Autos"
                ],
                "summary": "Delete user",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteCar",
                "parameters": [
                    {
                        "name": "Car id",
                        "in": "path",
                        "description": "Car id to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}